+++ NEWS.md
@@ -1,5 +1,9 @@
 # ebirdst 1.2021.1
 
+- fix bug causing species with same base code to be downloaded together, e.g. leafly also downloads leafly2, GitHub issue #43
+
+# ebirdst 1.2021.1
+
 - fix bug with extent in `load_fac_map_parameters()`, GitHub issue #40
 - use dynamic PAT cutoff in PPM calculations
 - update species list to account for second release of eBird data this year